+++ src/main/java/com/besson/arknights/block/custom/sideline/FiberCarpet.java
@@ -1,0 +1,33 @@
+package com.besson.arknights.block.custom.sideline;
+
+import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.util.function.BooleanBiFunction;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.util.shape.VoxelShapes;
+import net.minecraft.world.BlockView;
+
+import java.util.stream.Stream;
+
+public class FiberCarpet extends FurnitureHorizontalFacingBlock {
+    public FiberCarpet(Settings settings) {
+        super(settings);
+    }
+    private static final VoxelShape SHAPE_N = Block.createCuboidShape(-16, 0, 0, 16, 0.25, 32);
+    private static final VoxelShape SHAPE_W = Block.createCuboidShape(0, 0, 0, 32, 0.25, 32);
+    private static final VoxelShape SHAPE_S = Block.createCuboidShape(0, 0, -16, 32, 0.25, 16);
+    private static final VoxelShape SHAPE_E = Block.createCuboidShape(-16, 0, -16, 16, 0.25, 16);
+
+    @Override
+    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        return switch (state.get(FACING)){
+            case WEST -> SHAPE_W;
+            case SOUTH -> SHAPE_S;
+            case EAST -> SHAPE_E;
+            default -> SHAPE_N;
+        };
+    }
+}