+++ src/main/java/com/besson/arknights/block/custom/sideline/GrayishWall.java
@@ -1,0 +1,42 @@
+package com.besson.arknights.block.custom.sideline;
+
+import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import com.besson.arknights.block.ModPillarBlock;
+import net.minecraft.block.Block;
+import net.minecraft.block.BlockState;
+import net.minecraft.block.ShapeContext;
+import net.minecraft.item.ItemPlacementContext;
+import net.minecraft.state.StateManager;
+import net.minecraft.state.property.EnumProperty;
+import net.minecraft.util.StringIdentifiable;
+import net.minecraft.util.function.BooleanBiFunction;
+import net.minecraft.util.math.BlockPos;
+import net.minecraft.util.math.Direction;
+import net.minecraft.util.shape.VoxelShape;
+import net.minecraft.util.shape.VoxelShapes;
+import net.minecraft.world.BlockView;
+import net.minecraft.world.WorldAccess;
+import org.jetbrains.annotations.Nullable;
+
+import java.util.stream.Stream;
+
+public class GrayishWall extends FurnitureHorizontalFacingBlock {
+    private static final VoxelShape SHAPE_N = Block.createCuboidShape(-16, 0, 0, 16, 32, 16);
+    private static final VoxelShape SHAPE_W = Block.createCuboidShape(0, 0, 0, 16, 32, 32);
+    private static final VoxelShape SHAPE_S = Block.createCuboidShape(0, 0, 0, 32, 32, 16);
+    private static final VoxelShape SHAPE_E = Block.createCuboidShape(0, 0, -16, 16, 32, 16);
+
+    public GrayishWall(Settings settings) {
+        super(settings);
+    }
+
+    @Override
+    public VoxelShape getOutlineShape(BlockState state, BlockView world, BlockPos pos, ShapeContext context) {
+        return switch (state.get(FACING)){
+            case WEST -> SHAPE_W;
+            case SOUTH -> SHAPE_S;
+            case EAST -> SHAPE_E;
+            default -> SHAPE_N;
+        };
+    }
+}