+++ src/main/java/com/besson/arknights/block/custom/warehouse/ConcreteWall.java
@@ -1,6 +1,7 @@
 package com.besson.arknights.block.custom.warehouse;
 
 import com.besson.arknights.block.FurnitureHorizontalFacingBlock;
+import com.besson.arknights.block.ModDoubleBlock;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.ShapeContext;
@@ -8,7 +9,7 @@ import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.shape.VoxelShape;
 import net.minecraft.world.BlockView;
 
-public class ConcreteWall extends FurnitureHorizontalFacingBlock {
+public class ConcreteWall extends ModDoubleBlock {
     public static final VoxelShape SHAPE = Block.createCuboidShape(0,0,0,16,16,16);
     public ConcreteWall(Settings settings) {
         super(settings);