+++ src/main/java/com/besson/arknights/datagen/ModRecipesProvider.java
@@ -183,23 +183,23 @@ public class ModRecipesProvider extends FabricRecipeProvider {
                 .criterion("has_item", conditionsFromItem(Items.BLACK_DYE))
                 .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "black_nightstand"));
 
-        ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.SIMPLE_BLACK_DESK_LEFT, 1)
-                .pattern("###")
-                .pattern("A  ")
-                .pattern("###")
-                .input('#', ItemTags.PLANKS)
-                .input('A', Items.BLACK_DYE)
-                .criterion("has_item", conditionsFromItem(Items.BLACK_DYE))
-                .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "simple_black_desk_left"));
-
-        ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.SIMPLE_BLACK_DESK_RIGHT, 1)
-                .pattern("###")
-                .pattern("  A")
-                .pattern("###")
-                .input('#', ItemTags.PLANKS)
-                .input('A', Items.BLACK_DYE)
-                .criterion("has_item", conditionsFromItem(Items.BLACK_DYE))
-                .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "simple_black_desk_right"));
+//        ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.SIMPLE_BLACK_DESK_LEFT, 1)
+//                .pattern("###")
+//                .pattern("A  ")
+//                .pattern("###")
+//                .input('#', ItemTags.PLANKS)
+//                .input('A', Items.BLACK_DYE)
+//                .criterion("has_item", conditionsFromItem(Items.BLACK_DYE))
+//                .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "simple_black_desk_left"));
+//
+//        ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.SIMPLE_BLACK_DESK_RIGHT, 1)
+//                .pattern("###")
+//                .pattern("  A")
+//                .pattern("###")
+//                .input('#', ItemTags.PLANKS)
+//                .input('A', Items.BLACK_DYE)
+//                .criterion("has_item", conditionsFromItem(Items.BLACK_DYE))
+//                .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "simple_black_desk_right"));
 
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.BLACK_OFFICE_CHAIR, 1)
                 .pattern("A A")
@@ -219,33 +219,33 @@ public class ModRecipesProvider extends FabricRecipeProvider {
                 .criterion("has_item", conditionsFromItem(Items.BLACK_DYE))
                 .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "simple_black_cabinet"));
 
-        ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.CHECKERBOARD_BOOKCASE1, 1)
-                .pattern("###")
-                .pattern("AAA")
-                .input('#', ItemTags.PLANKS)
-                .input('A', Items.BLACK_DYE)
-                .criterion("has_item", conditionsFromItem(Items.BLACK_DYE))
-                .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "checkerboard_bookcase1"));
-
-        ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.CHECKERBOARD_BOOKCASE2, 1)
-                .pattern("###")
-                .pattern("ABA")
-                .pattern("###")
-                .input('#', ItemTags.PLANKS)
-                .input('A', Items.WHITE_DYE)
-                .input('B', Items.BLACK_DYE)
-                .criterion("has_item", conditionsFromItem(Items.WHITE_DYE))
-                .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "checkerboard_bookcase2"));
-
-        ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.CHECKERBOARD_BOOKCASE3, 1)
-                .pattern("###")
-                .pattern("ABA")
-                .pattern("###")
-                .input('#', ItemTags.PLANKS)
-                .input('A', Items.WHITE_DYE)
-                .input('B', Items.BLACK_DYE)
-                .criterion("has_item", conditionsFromItem(Items.WHITE_DYE))
-                .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "checkerboard_bookcase3"));
+//        ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.CHECKERBOARD_BOOKCASE1, 1)
+//                .pattern("###")
+//                .pattern("AAA")
+//                .input('#', ItemTags.PLANKS)
+//                .input('A', Items.BLACK_DYE)
+//                .criterion("has_item", conditionsFromItem(Items.BLACK_DYE))
+//                .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "checkerboard_bookcase1"));
+//
+//        ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.CHECKERBOARD_BOOKCASE2, 1)
+//                .pattern("###")
+//                .pattern("ABA")
+//                .pattern("###")
+//                .input('#', ItemTags.PLANKS)
+//                .input('A', Items.WHITE_DYE)
+//                .input('B', Items.BLACK_DYE)
+//                .criterion("has_item", conditionsFromItem(Items.WHITE_DYE))
+//                .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "checkerboard_bookcase2"));
+//
+//        ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.CHECKERBOARD_BOOKCASE3, 1)
+//                .pattern("###")
+//                .pattern("ABA")
+//                .pattern("###")
+//                .input('#', ItemTags.PLANKS)
+//                .input('A', Items.WHITE_DYE)
+//                .input('B', Items.BLACK_DYE)
+//                .criterion("has_item", conditionsFromItem(Items.WHITE_DYE))
+//                .offerTo(exporter, new Identifier(ArknightsFurniture.MOD_ID, "checkerboard_bookcase3"));
 
         ShapedRecipeJsonBuilder.create(RecipeCategory.BUILDING_BLOCKS, ModBlocks.BLACK_HD_TV, 1)
                 .pattern("#BA")
@@ -2718,13 +2718,8 @@ public class ModRecipesProvider extends FabricRecipeProvider {
 
         registerFurnParts(exporter, ModBlocks.BLACK_BED, 125);
         registerFurnParts(exporter, ModBlocks.BLACK_NIGHTSTAND, 48);
-        registerFurnParts(exporter, ModBlocks.SIMPLE_BLACK_DESK_LEFT, 41);
-        registerFurnParts(exporter, ModBlocks.SIMPLE_BLACK_DESK_RIGHT, 42);
         registerFurnParts(exporter, ModBlocks.BLACK_OFFICE_CHAIR, 55);
         registerFurnParts(exporter, ModBlocks.SIMPLE_BLACK_CABINET, 48);
-        registerFurnParts(exporter, ModBlocks.CHECKERBOARD_BOOKCASE1, 33);
-        registerFurnParts(exporter, ModBlocks.CHECKERBOARD_BOOKCASE2, 33);
-        registerFurnParts(exporter, ModBlocks.CHECKERBOARD_BOOKCASE3, 33);
         registerFurnParts(exporter, ModBlocks.BLACK_HD_TV, 75);
         registerFurnParts(exporter, ModBlocks.BLACK_AND_WHITE_SQUARE_TABLE, 70);
         registerFurnParts(exporter, ModBlocks.LOW_BLACK_AND_WHITE_SOFA, 83);